chore: JobRun 도메인 세팅
